discord/pyproject.toml
[tool.poetry]
name = "fastapi-app"
version = "0.1.0"
description = "RAG_endpoint"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.2"
fastapi = "^0.115.7"
uvicorn = "^0.34.0"
click = "^8.1.8"
python-dotenv = "^1.0.1"
langchain = "^0.3.16"
langchain-community = "^0.3.16"
transformers = "^4.48.1"
pillow = "^11.1.0"
chromadb = "^0.6.3"
langchain-openai = "^0.3.3"
langgraph = "^0.2.69"
streamlit = "^1.42.0"
protobuf = "^5.29.3"
watchdog = "^6.0.0"
umap-learn = "^0.5.3"
plotly = "^5.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
